Making adjustments in TestImageLandsat5, TestImageHelper and SEBAL
Some adjustments in these three classes were necessary to obtain Rho
correctly. One of these adjustments was the change of return type of
method F1 to List<ImagePixel>, this way, was possible to use
ProcessPixel in setInitialProperties instead of processPixelQuenteFrio
in F1. With this change, Rho was calculated in ProcessPixel and stored
in an output that was further stored in the actual list of pixels
compared in the asserts with the expected list of pixels.